backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch==2.2.2
torchvision==0.17.2
Pillow==10.2.0
